Remove duplicate TargetType enum from verify_config
